third_party/blink/web_tests/external/wpt/css/css-animations/parsing/animation-name-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['animation-name'] = "\\"something\\"" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['animation-name'] = "\\"---\\\\22---\\"" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['animation-name'] = "\\"multi word string\\"" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['animation-name'] = "\\"none\\"" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['animation-name'] = "\\"none\\", \\"INITIAL\\", \\"inherit\\"" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['animation-name'] = "\\"none\\", both, ease-in" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['animation-name'] = "\\"NoNe\\"" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['animation-name'] = "\\"initial\\"" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['animation-name'] = "\\"INITIAL\\"" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['animation-name'] = "\\"inherit\\"" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['animation-name'] = "\\"INHERIT\\"" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['animation-name'] = "\\"revert\\"" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['animation-name'] = "\\"REVERT\\"" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['animation-name'] = "\\"revert-layer\\"" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['animation-name'] = "\\"REVERT-LAYER\\"" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['animation-name'] = "\\"unset\\"" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['animation-name'] = "\\"UNSET\\"" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['animation-name'] = "\\"default\\"" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['animation-name'] = "\\"DEFAULT\\"" should set the property value
  assert_not_equals: property should be set got disallowed value ""
Harness: the test ran to completion.
